+++ tradingbot/__init__.py
@@ -8,18 +8,13 @@ from tradingbot.strategies.strategy import Strategy
 from tradingbot.strategies.basic_strategy import BasicStrategy
 
 from tradingbot.config import Config
-from tradingbot.files import (
-    Files, try_load_json, try_read_file, lock, unlock,
-    reset_successful_symbols_file, reset_consecutive_times_down_file,
-    get_last_balance, get_consecutive_times_down,
-    increment_consecutive_times_down, write_file, try_remove_file
-)
 from tradingbot.log import log
 from tradingbot.mt_client import MT_Client
 from tradingbot.utils import (
-    string_to_date_utc,
-    get_remaining_symbols,
-    add_successful_symbol
+    string_to_date_utc, create_magic_number, reset_successful_symbols,
+    add_successful_symbol, reboot_mt, reset_consecutive_times_down_file,
+    get_successful_symbols, get_remaining_symbols, get_last_balance,
+    get_consecutive_times_down, increment_consecutive_times_down
 )
 from tradingbot.ohlc import OHLC
 from tradingbot.order_operations import OrderOperations
@@ -27,6 +22,7 @@ from tradingbot.order_type import OrderType
 from tradingbot.order import (
     Order, MutableOrderDetails, ImmutableOrderDetails, OrderPrice
 )
+from tradingbot.context_managers.blocker import Blocker
 from tradingbot.trading_methods import (get_pip, get_pivots, EMA, RSI, SAR,
                                         confirmation_pattern,
                                         three_bar_reversal, pinbar_pattern,